circos.ideogram {GWLD}	R Documentation

circos snp

Description

circos snp

Usage

circos.ideogram(
  data,
  ylim = c(0, 1),
  cell.padding = c(0.02, 0, 0.02, 0),
  track.height = 0.1,
  cex.lab = 0.6,
  bg.border = NA,
  family = NULL,
  ...
)

Arguments

data
data.frame

ylim
Range of data on y-axis

cell.padding
only affect current track

track.height
Height of the track. It is the percentage to the radius of the unit circles

cex.lab
a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the defaul

bg.border
Color for the border of the plotting regions

family
the name of a font family for drawing text

...
other graphical parameters

Examples

NULL

[Package GWLD version 1.3.4 Index]